Struct inexor_rgf_rt::runtime::builder::RuntimeBuilder

pub struct RuntimeBuilder<L, R> { /* private fields */ }

Implementations

impl RuntimeBuilder<SetConfigLocations, NotRunning>

pub fn new() -> Self

pub fn instance_config<P: Into<OptionOption<PathBuf>>>( self, location: P ) -> RuntimeBuilder<SetConfigLocations, NotRunning>

Sets the location of the instance configuration.

pub fn graphql_server_config<P: Into<OptionOption<PathBuf>>>( self, location: P ) -> RuntimeBuilder<SetConfigLocations, NotRunning>

Sets the location of the graphql server configuration.

pub fn plugins_config<P: Into<OptionOption<PathBuf>>>( self, location: P ) -> RuntimeBuilder<SetConfigLocations, NotRunning>

Sets the location of the plugins configuration.

pub async fn load_config_files( self ) -> RuntimeBuilder<ConfigFilesLoaded, NotRunning>

Loads the config files and transits to state ConfigFilesLoaded.

pub fn ignore_config_files( self ) -> RuntimeBuilder<ConfigFilesLoaded, NotRunning>

Doesn’t load the config files. Transits to state ConfigFilesLoaded.

impl RuntimeBuilder<ConfigFilesLoaded, NotRunning>

pub fn instance_name<S: Into<OptionOption<String>>>( self, name: S ) -> RuntimeBuilder<ConfigFilesLoaded, NotRunning>

Sets the name of the instance.

pub fn instance_description<S: Into<OptionOption<String>>>( self, description: S ) -> RuntimeBuilder<ConfigFilesLoaded, NotRunning>

Sets the description of the instance.

pub fn hostname<S: Into<OptionOption<String>>>( self, hostname: S ) -> RuntimeBuilder<ConfigFilesLoaded, NotRunning>

Sets the hostname of the GraphQL server.

pub fn port<S: Into<OptionOption<u16>>>( self, port: S ) -> RuntimeBuilder<ConfigFilesLoaded, NotRunning>

Sets the port number of the GraphQL server.

pub fn pick_free_port(self) -> RuntimeBuilder<ConfigFilesLoaded, NotRunning>

Picks a free port instead of a fixed port number.

pub fn secure<S: Into<OptionOption<bool>>>( self, secure: S ) -> RuntimeBuilder<ConfigFilesLoaded, NotRunning>

Sets if the GraphQL server shall use https or http.

pub fn address( self, address: &InstanceAddress ) -> RuntimeBuilder<ConfigFilesLoaded, NotRunning>

Sets the hostname, port and secure of the GraphQL server.

pub fn shutdown_timeout<S: Into<OptionOption<u64>>>( self, shutdown_timeout: S ) -> RuntimeBuilder<ConfigFilesLoaded, NotRunning>

Sets timeout of the shutdown of the GraphQL server.

pub fn workers<S: Into<OptionOption<usize>>>( self, workers: S ) -> RuntimeBuilder<ConfigFilesLoaded, NotRunning>

Sets the number of workers of the GraphQL server.

pub fn default_context_path<S: Into<OptionOption<String>>>( self, default_context_path: S ) -> RuntimeBuilder<ConfigFilesLoaded, NotRunning>

Sets context path of a web resource provider which shall be used as default context path.

pub fn disable_all_plugins<S: Into<OptionOption<bool>>>( self, disabled: S ) -> RuntimeBuilder<ConfigFilesLoaded, NotRunning>

Disables all plugins.

pub fn disabled_plugins<S: Into<OptionOption<Vec<String>>>>( self, disabled_plugins: S ) -> RuntimeBuilder<ConfigFilesLoaded, NotRunning>

Sets which plugins will be disabled.

pub fn disable_hot_deploy<S: Into<OptionOption<bool>>>( self, disabled: S ) -> RuntimeBuilder<ConfigFilesLoaded, NotRunning>

Disables hot deployment of plugins.

pub fn hot_deploy_location<S: Into<OptionOption<String>>>( self, hot_deploy_location: S ) -> RuntimeBuilder<ConfigFilesLoaded, NotRunning>

Sets the directory where plugins can be hot deployed.

pub fn install_location<S: Into<OptionOption<String>>>( self, install_location: S ) -> RuntimeBuilder<ConfigFilesLoaded, NotRunning>

Sets the directory where plugins are located. During hot deployment new plugins will be moved into this directory.

pub fn get(self) -> Arc<dyn Runtime>

pub async fn init(self) -> RuntimeBuilder<ConfigFilesLoaded, Initialized>

pub async fn block_on(self) -> RuntimeBuilder<ConfigFilesLoaded, Shutdown>

impl RuntimeBuilder<ConfigFilesLoaded, Initialized>

pub async fn post_init(self) -> RuntimeBuilder<ConfigFilesLoaded, Ready>

impl RuntimeBuilder<ConfigFilesLoaded, Ready>

pub fn get(self) -> Arc<dyn Runtime>

pub async fn with_runtime<F, C>( self, f: C ) -> RuntimeBuilder<ConfigFilesLoaded, Ready>where F: Future<Output = ()>, C: FnOnce(Arc<dyn Runtime>) -> F,

pub async fn spawn(self) -> RuntimeBuilder<ConfigFilesLoaded, Running>

Starts the GraphQL server (non-blocking).

pub async fn spawn_blocking(self) -> RuntimeBuilder<ConfigFilesLoaded, Finished>

Starts the GraphQL server (blocking).

pub async fn run_for( self, duration: Duration ) -> RuntimeBuilder<ConfigFilesLoaded, Finished>

Runs starts the GraphQL server. Stops the GraphQL server after the given duration.

pub async fn do_not_run(self) -> RuntimeBuilder<ConfigFilesLoaded, Finished>

Do not start the GraphQL server but shutdown the runtime.

impl RuntimeBuilder<ConfigFilesLoaded, Running>

pub fn get(self) -> Arc<dyn Runtime>

pub async fn with_runtime<F, C>( self, f: C ) -> RuntimeBuilder<ConfigFilesLoaded, Running>where F: Future<Output = ()>, C: FnOnce(Arc<dyn Runtime>) -> F,

pub async fn stop(self) -> RuntimeBuilder<ConfigFilesLoaded, Finished>

Stops the runtime. Waits for the GraphQL server has been stopped.

pub async fn stop_with_timeout( self, timeout_duration: Duration ) -> RuntimeBuilder<ConfigFilesLoaded, Finished>

Stops the runtime. Waits for the GraphQL server has been stopped or the given timeout has been reached.

pub async fn wait_for_stopped( self ) -> RuntimeBuilder<ConfigFilesLoaded, Finished>

Waits for the GraphQL server has been stopped.

pub async fn wait_for_stopped_with_timeout( self, timeout_duration: Duration ) -> RuntimeBuilder<ConfigFilesLoaded, Finished>

Waits for the GraphQL server has been stopped.

impl RuntimeBuilder<ConfigFilesLoaded, Finished>

pub fn get(self) -> Arc<dyn Runtime>

pub async fn with_runtime<F, C>( self, f: C ) -> RuntimeBuilder<ConfigFilesLoaded, Finished>where F: Future<Output = ()>, C: FnOnce(Arc<dyn Runtime>) -> F,

pub async fn pre_shutdown( self ) -> RuntimeBuilder<ConfigFilesLoaded, PreShutdown>

impl RuntimeBuilder<ConfigFilesLoaded, PreShutdown>

pub async fn shutdown(self) -> RuntimeBuilder<ConfigFilesLoaded, Shutdown>

impl RuntimeBuilder<ConfigFilesLoaded, Shutdown>

pub async fn wait_for( self, duration: Duration ) -> RuntimeBuilder<ConfigFilesLoaded, Shutdown>