Trait inexor_rgf_rt::runtime::runtime::Runtime

pub trait Runtime: Send + Sync {

    // Required methods
    fn config<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn init<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn post_init<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn pre_shutdown<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn shutdown<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn run<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn stop(&self);
    fn is_running(&self) -> bool;
    fn wait_for_started<'life0, 'async_trait>(
        &'life0 self,
        timeout_duration: Duration
    ) -> Pin<Box<dyn Future<Output = Result<(), Elapsed>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn wait_for_stopped<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn wait_for_stopped_with_timeout<'life0, 'async_trait>(
        &'life0 self,
        timeout_duration: Duration
    ) -> Pin<Box<dyn Future<Output = Result<(), Elapsed>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_command_manager(&self) -> Arc<dyn CommandManager>;
    fn get_component_manager(&self) -> Arc<dyn ComponentManager>;
    fn get_config_manager(&self) -> Arc<dyn ConfigManager>;
    fn get_dynamic_graph_query_service(
        &self
    ) -> Arc<dyn DynamicGraphQueryService>;
    fn get_dynamic_graph_schema_manager(
        &self
    ) -> Arc<dyn DynamicGraphSchemaManager>;
    fn get_entity_behaviour_manager(&self) -> Arc<dyn EntityBehaviourManager>;
    fn get_entity_behaviour_registry(&self) -> Arc<dyn EntityBehaviourRegistry>;
    fn get_entity_component_behaviour_manager(
        &self
    ) -> Arc<dyn EntityComponentBehaviourManager>;
    fn get_entity_component_behaviour_registry(
        &self
    ) -> Arc<dyn EntityComponentBehaviourRegistry>;
    fn get_entity_instance_manager(&self) -> Arc<dyn EntityInstanceManager>;
    fn get_entity_type_manager(&self) -> Arc<dyn EntityTypeManager>;
    fn get_entity_vertex_manager(&self) -> Arc<dyn EntityVertexManager>;
    fn get_event_manager(&self) -> Arc<dyn SystemEventManager>;
    fn get_flow_instance_manager(&self) -> Arc<dyn FlowInstanceManager>;
    fn get_graph_database(&self) -> Arc<dyn GraphDatabase>;
    fn get_graphql_query_service(&self) -> Arc<dyn GraphQLQueryService>;
    fn get_graphql_schema_manager(&self) -> Arc<dyn GraphQLSchemaManager>;
    fn get_graphql_server(&self) -> Arc<dyn GraphQLServer>;
    fn get_instance_service(&self) -> Arc<dyn InstanceService>;
    fn get_namespace_manager(&self) -> Arc<dyn NamespaceManager>;
    fn get_plugin_container_manager(&self) -> Arc<dyn PluginContainerManager>;
    fn get_plugin_repository_manager(&self) -> Arc<dyn PluginRepositoryManager>;
    fn get_reactive_entity_instance_manager(
        &self
    ) -> Arc<dyn ReactiveEntityInstanceManager>;
    fn get_reactive_flow_instance_manager(
        &self
    ) -> Arc<dyn ReactiveFlowInstanceManager>;
    fn get_reactive_relation_instance_manager(
        &self
    ) -> Arc<dyn ReactiveRelationInstanceManager>;
    fn get_relation_behaviour_manager(
        &self
    ) -> Arc<dyn RelationBehaviourManager>;
    fn get_relation_behaviour_registry(
        &self
    ) -> Arc<dyn RelationBehaviourRegistry>;
    fn get_relation_component_behaviour_manager(
        &self
    ) -> Arc<dyn RelationComponentBehaviourManager>;
    fn get_relation_component_behaviour_registry(
        &self
    ) -> Arc<dyn RelationComponentBehaviourRegistry>;
    fn get_relation_edge_manager(&self) -> Arc<dyn RelationEdgeManager>;
    fn get_relation_instance_manager(&self) -> Arc<dyn RelationInstanceManager>;
    fn get_relation_type_manager(&self) -> Arc<dyn RelationTypeManager>;
    fn get_remotes_manager(&self) -> Arc<dyn RemotesManager>;
    fn get_shutdown_manager(&self) -> Arc<dyn ShutdownManager>;
    fn get_web_resource_manager(&self) -> Arc<dyn WebResourceManager>;

}

Required Methods

fn config<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn init<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn post_init<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn pre_shutdown<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn shutdown<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn run<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn stop(&self)

fn is_running(&self) -> bool

fn wait_for_started<'life0, 'async_trait>( &'life0 self, timeout_duration: Duration ) -> Pin<Box<dyn Future<Output = Result<(), Elapsed>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Waits for the GraphQL server to be started. Times out if the GraphQL server is not running after the given duration.

fn wait_for_stopped<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Waits for the GraphQL server has been stopped.

fn wait_for_stopped_with_timeout<'life0, 'async_trait>( &'life0 self, timeout_duration: Duration ) -> Pin<Box<dyn Future<Output = Result<(), Elapsed>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Waits for the GraphQL server has been stopped. Times out if the GraphQL server is still running after the given duration.

fn get_command_manager(&self) -> Arc<dyn CommandManager>

fn get_component_manager(&self) -> Arc<dyn ComponentManager>

fn get_config_manager(&self) -> Arc<dyn ConfigManager>

fn get_dynamic_graph_query_service(&self) -> Arc<dyn DynamicGraphQueryService>

fn get_dynamic_graph_schema_manager(&self) -> Arc<dyn DynamicGraphSchemaManager>

fn get_entity_behaviour_manager(&self) -> Arc<dyn EntityBehaviourManager>

fn get_entity_behaviour_registry(&self) -> Arc<dyn EntityBehaviourRegistry>

fn get_entity_component_behaviour_manager( &self ) -> Arc<dyn EntityComponentBehaviourManager>

fn get_entity_component_behaviour_registry( &self ) -> Arc<dyn EntityComponentBehaviourRegistry>

fn get_entity_instance_manager(&self) -> Arc<dyn EntityInstanceManager>

fn get_entity_type_manager(&self) -> Arc<dyn EntityTypeManager>

fn get_entity_vertex_manager(&self) -> Arc<dyn EntityVertexManager>

fn get_event_manager(&self) -> Arc<dyn SystemEventManager>

fn get_flow_instance_manager(&self) -> Arc<dyn FlowInstanceManager>

fn get_graph_database(&self) -> Arc<dyn GraphDatabase>

fn get_graphql_query_service(&self) -> Arc<dyn GraphQLQueryService>

fn get_graphql_schema_manager(&self) -> Arc<dyn GraphQLSchemaManager>

fn get_graphql_server(&self) -> Arc<dyn GraphQLServer>

fn get_instance_service(&self) -> Arc<dyn InstanceService>

fn get_namespace_manager(&self) -> Arc<dyn NamespaceManager>

fn get_plugin_container_manager(&self) -> Arc<dyn PluginContainerManager>

fn get_plugin_repository_manager(&self) -> Arc<dyn PluginRepositoryManager>

fn get_reactive_entity_instance_manager( &self ) -> Arc<dyn ReactiveEntityInstanceManager>

fn get_reactive_flow_instance_manager( &self ) -> Arc<dyn ReactiveFlowInstanceManager>

fn get_reactive_relation_instance_manager( &self ) -> Arc<dyn ReactiveRelationInstanceManager>

fn get_relation_behaviour_manager(&self) -> Arc<dyn RelationBehaviourManager>

fn get_relation_behaviour_registry(&self) -> Arc<dyn RelationBehaviourRegistry>

fn get_relation_component_behaviour_manager( &self ) -> Arc<dyn RelationComponentBehaviourManager>

fn get_relation_component_behaviour_registry( &self ) -> Arc<dyn RelationComponentBehaviourRegistry>

fn get_relation_edge_manager(&self) -> Arc<dyn RelationEdgeManager>

fn get_relation_instance_manager(&self) -> Arc<dyn RelationInstanceManager>

fn get_relation_type_manager(&self) -> Arc<dyn RelationTypeManager>

fn get_remotes_manager(&self) -> Arc<dyn RemotesManager>

fn get_shutdown_manager(&self) -> Arc<dyn ShutdownManager>

fn get_web_resource_manager(&self) -> Arc<dyn WebResourceManager>